+++ plugin_tooling/src/com/github/rustdt/tooling/ops/RacerCompletionOperation.java
@@ -15,6 +15,7 @@ import melnorme.lang.tooling.ops.IProcessRunner;
 import melnorme.utilbox.collections.ArrayList2;
 import melnorme.utilbox.concurrency.OperationCancellation;
 import melnorme.utilbox.core.CommonException;
+import melnorme.utilbox.misc.Location;
 import melnorme.utilbox.process.ExternalProcessHelper.ExternalProcessResult;
 
 public class RacerCompletionOperation extends RacerOperation {
@@ -22,9 +23,18 @@ public class RacerCompletionOperation extends RacerOperation {
 	protected final int offset;
 	
 	public RacerCompletionOperation(IProcessRunner processRunner, String toolPath, String rustPath, int offset,
-			ArrayList2<String> arguments) {
-		super(processRunner, toolPath, rustPath, arguments);
+			int line_0, int col_0, Location fileLocation) {
+		super(processRunner, toolPath, rustPath, getArguments(line_0, col_0, fileLocation));
+		
 		this.offset = offset;
+	}
+	
+	protected static ArrayList2<String> getArguments(int line_0, int col_0, Location fileLocation) {
+		ArrayList2<String> arguments = new ArrayList2<String>("complete-with-snippet");
+		arguments.add(Integer.toString(line_0 + 1)); // use one-based index
+		arguments.add(Integer.toString(col_0)); // But this one is zero-based index
+		arguments.add(fileLocation.toPathString());
+		return arguments;
 	}
 	
 	public LangCompletionResult executeAndProcessOutput() throws CommonException, OperationCancellation {